+++ app/irsystem/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+# Define a Blueprint for this module (mchat)
+irsystem = Blueprint('irsystem', __name__, url_prefix='/', static_folder='static', template_folder='templates')
+
+# Import all controllers
+# from .controllers.search_controller import *
