+++ app/irsystem/controllers/__init__.py
@@ -1,0 +1,28 @@
+# Import flask deps
+from flask import request, render_template, \
+    flash, g, session, redirect, url_for, jsonify, abort
+
+# For decorators around routes
+from functools import wraps
+
+# Import for pass / encryption
+from werkzeug import check_password_hash, generate_password_hash
+
+# Import the db object from main app module
+from app import db
+
+# Marshmallow
+from marshmallow import ValidationError
+
+# Import socketio for socket creation in this module
+from app import socketio
+
+# Import module models
+# from app.irsystem import search
+
+# IMPORT THE BLUEPRINT APP OBJECT
+from app.irsystem import irsystem
+
+# Import module models
+from app.accounts.models.user import *
+from app.accounts.models.session import *
