+++ app/irsystem/controllers/search_controller.py
@@ -1,6 +1,6 @@
-from . import *  
-from api.irsystem.models.helpers import *
-from api.irsystem.models.helpers import NumpyEncoder as NumpyEncoder
+# from . import *
+# from app.irsystem.models.helpers import *
+# from app.irsystem.models.helpers import NumpyEncoder as NumpyEncoder
 
 # project_name = "ClubHub"
 # net_id = "Cora Wu (cjw322), Jonathan Gao (jg992), Josiah Kek (jrk322), Rishabh Sarup (rs868), Samuel Lye (sl2982)"
@@ -15,6 +15,3 @@ from api.irsystem.models.helpers import NumpyEncoder as NumpyEncoder
 # 		output_message = "Your search: " + query
 # 		data = range(5)
 # 	return render_template('search.html', name=project_name, netid=net_id, output_message=output_message, data=data)
-
-
-
