+++ requirements.txt
@@ -54,6 +54,5 @@ six==1.14.0
 SQLAlchemy==1.3.13
 toml==0.10.2
 urllib3==1.25.8
-waitress==2.0.0
 Werkzeug==0.16.1
 WTForms==2.2.1
